Move OAuth error handling to runtime layer
Move OAuthAuthorizationRequiredError and ServerInfoProvider from agent
to runtime package, centralizing OAuth logic where flows are orchestrated.
All agent.Tools() calls now use runtime helper for consistent error handling.
